Add --library option for tracing library functions
For tracing library functions, it needs to know the actual loaded
address of the library.  So record the /proc/self/maps data before
recording rstacks, but as it alters the format of the ftrace.data file
bump up the file version too.

For now, this --library option only works with --use-pipe.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>