Add support tracing multi-process application
Some processes fork them self to work parellely, but it doesn't
be supported by ftrace - it uses stdio stream for recording with
full-buffering.

But since the fork() cannot know about the buffer, it didn't flush
the buffer so it ended up being corrupted.  Use --use-pipe option
will support such an application as it doesn't use the stdio stream.

So now we can remove the checking of tid when recording rstack
because it now can be called after fork().

Signed-off-by: Namhyung Kim <namhyung@gmail.com>