Fix occasional bug for t-abcd and t-fork testcases
There was a time window that check_tid_list() scans all process/thread
and then reads remaining FORK_START message.  In this case it'll see a
tid of -1 in the tid_list resulted in failure at replay time.

Check the tid list again after reading remaining messages from pipe and,
if so, wait for next FORK_END message to get a pid.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>