test: Implement new test framework
The runtest.py script is to test various cases flexibly.  It requires
an argument that is a name or number of the testcase.  Passing 'all'
runs test for all cases.  There should be a test script like
t<number_<name>.py for each testcase to be run.  The test script
should define a new TestCase class which inherits TestBase class.

Normally it's enough just to define test name and the result.  The
TestBase class takes care of building, running and comparing output.
Each test case can extend runcmd() and sort() method (or even build()
or run() methods if needed) to do what the test wants.

The test will be done for each -pg and -finstrument-functions and all
optimization levels.  A testcase is considered 'OK' if the output of
ftrace is exactly match to expected result (after sort() of course)
and the exitcode is zero.

Each two-character result has following meaning:

  OK: Test succeeded
  LA: Unsupported Language
  BI: Build failed
  SG: Abnormal exit by signal
  TM: Test ran too long
  NG: Different test result
  NZ: Non-zero return value

Signed-off-by: Namhyung Kim <namhyung@gmail.com>