test: Update vfork & exec testcase
This test is to check tracing output correct if the test program
calling vfork() and exec() to create a child process to execute some
other program.  The vfork() is still used for a compatibility reason I
guess.  For simplicity, it uses t-abc program for new child and it
doesn't enable tracing.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>