runtest.py: add ignore_children argument to sort()
When a test program calls one of exec() family function, it starts to
execute another program and ftrace has no idea of its symbols then.
Also the exec-ed program may not have tracing enabled.  So comparing
output of the child is almost meaningless.

So make runtest.py to allow to ignore the child's result.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>