mcount: Introduce struct ftrace_ret_stack
It is to be used to save function info in more compact format using
bit fields.  Most of mcount_ret_stack info will be there.  Also add
->type field to distinguish function entry or exit more easily.

Now addr field is saved in 48 bit and remaining 16 bit is used for
other purpose.  But IMHO it's not a problem since all 64 bit systems
it supports only uses 48 bit for actual addresses.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>