mcount/x86: Save and restore rdx register in mcount_return()
I found that node.js accesses the rdx register to compare with a value
right after calling a function.  AFAIK it's a caller-saved register so
that it can be changed during function calls.  To ensure or access the
register, the calling function must save its original value before
calling functions.

However the rdx register can be used to carry return value of some
instructions like imul or idiv.  So it anyway needs to be preserved.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>