mcount: Introduce mcount_restore() and mcount_reset()
This is to be used by some programs which deal with stack frames
directly.  For example, node.js garbage collector iterates frames and
try to move some code objects.  In that case, it'll see invalid
address and crashed.

To avoid that, mcount_restore() can be used to restore the original
return addresses() and it must call mcount_reset() in the same
function after accessing stack frames.  Of course, this requires to
link the binary with libmcount.so library.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>