ftrace: Use task->tid instead of rstack->tid
As rstack is maintained for each task, all tid values are same for a
single task.  So there's no reason to carry the tid in rstack.  This
is a preparation to get rid of it.  Just use task->tid instead of
rstack->tid.  To do that, pass ftrace_task_handle directly rather than
mcount_ret_stack when dealing with rstack.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>