ftrace: Use rstack->depth instead of task->stack_count
The stack_count variable is to keep task state and print remaining
functions at the end (if exists).  However it assumes that the rstack
entries always come with correct order without lost.  It might be not
true in some situation so use rstack->depth instead so that we can
proceed in case of the lost.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>