record: Unlink shmem buffer at the end
Now it reuses shmem buffers for double buffering, we cannot unlink the
shmem after writing to file.  Instead keep a list of shmem buffer and
unlink at the end of ftrace record.  To do that, adds a flag field in
the shmem buffer header and uses the SHMEM_FL_NEW bit.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>