ftrace: Fix bug in plthook/session bit check
When we ran ftrace record with --no-plthook, it'd complain about map
file at replay/report time like below:

  $ ftrace record -f xxx --no-plthook bench-cyg
  $ ftrace replay -f xxx
  ftrace: ERROR: ftrace.c:1525:read_map_file:
                 cannot open maps file: No such file or directory
  $ ls xxx
  4846.dat  info  sid-d7994f095256fef5.map  task

This is because of an error which checks wrong bit mask in
open_data_file().  The feat_mask checks TASK_SESSION but it actually a
bit index not a bitmask.  So it needs to be changed.  However, I also
think that it's a common mistake so it'd be better to change the
definition of the bitmask so that it can prevent possible similar
mistasks.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>