ftrace: Use libmcount-fast-single.so for single-thread apps
When no thread is used in the application, there's no reason to use
pthread API in the mcount() too.  So build a separate library for
single-threaded apps to speed up the processing ~20%.

Note that libpthread.so is still needed to be linked since we want to
use pthread_atfork() handler anyway.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>