mcount: Add basic support for setjmp/longjmp
Those functions are used to implement non-local goto, so it's hard to
trace the control flow after longjmp().  This patch saves a recent
invocation of setjmp() so that the later longjmp() can be redirected to
there in ftrace.

Note that it currently supports a single setjmp/longjmp only.  To
support multiple pairs, it should access the jmp_buf.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>