mcount: Fix a bug in cygprof entry/exit
It was assumed that the parent and child given to __cyg_profile_func_
enter/exit() are always same.  So the ftrace checks the address
between entry and exit to determine whether it was filtered or not.
But in some cases, it seems not true anymore.  Use an explicit flag to
specify a filtered function instead.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>