mcount: Fix filtering in cygprof functions
Filtering in cygprof differs from mcount since the exit path always
called even for filtered functions.  Also intermediate filters should
not interfere the output (depth).  So this patch adds
mcount_record_idx to keep the correct index with filtering enabled.

For mcount logic, it's same as mcount_rstack_idx, but for cygprof it
can be different.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>