ftrace: Add/Rename -T option for trigger w/o filters
From now on, the -T option will be used to setup triggers.  The syntax
is same as filter triggers but it doesn't setup filters.  So it can add
triggers to any function without disturbing filter state.

For example, setting a trigger with filter affects to not record sibling
functions, while a trigger *without* filter doesn't.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>