mcount: Fix PLT record after filtering
If a PLT entry was filtered, later invocation of the function wasn't
recorded since it didn't hook the return address.  So the PLT index was
ovewritten by the resolver function and then the plthook_entry() never
get called again.

Fix it by hooking the return address always and mark the rstack as
NORECORD.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>