test: Add fixup() function to handle known differences
In some C++ programs, constructors are often omitted when compiled with
optimization levels (and -pg option).  So the output is different to
expected result in this case.  Add fixup() function to TestCase class so
that it can tweak the result to be matched.

The TEST_SUCCESS_FIXED return value was added and prints 'OK' in yellow.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>