build: Add 'make config' target and configure script
The configure script is to save various config settings.  It was
parted done by FLAGS file, but it required same setting should be give
to make everytime.

By using this 'make config' (or ./configure directly), users can set
those config variables once and use them everytime.  The config
settings are saved to .config file in the current directory.  This
will be included by Makefile and override others.  If you want to
change the settings, you can rerun the make config with new settings
or simply delete the .config file (it'll run 'make config' next time
automatically).

For convenience, 'make config' will be called automatically when
there's no .config file in the current directory.  The settings can be
given using shell variables or environment variables.

For example, it'll install ftrace to /usr/bin/ correctly.

  $ make prefix=/usr
  $ sudo make install

This is actually same as below:

  $ make config prefix=/usr  (or ./configure --prefix=/usr)
  $ make
  $ sudo make install

Signed-off-by: Namhyung Kim <namhyung@gmail.com>