mcount: Fix PLT recursion with filter
We should set recursion guard when (symbol or depth) filter is
applied.  Otherwise, the plthook_saved_addr might be overwritten.
This probably lose traces of some PLT functions but an overwritten PLT
address caused a segfault during a test run of v8.

This patch also contains an unrelated whitespace fix.

Reported-by: Honggyu Kim <hong.gyu.kim@lge.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>