mcount: Fix shmem buffer cleanup
It missed to set TSD when allocating the buffers so destructor was not
called during thread termination.  The end result was that there're many
shmem buffers remaining in the /dev/shm directory after recording.

Set the TSD with shmem_curr and move prepare_shmem_buffer() after
mcount_init_file() so that the TSD key was initialized first.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>