mcount: Fix segfault on ARM when using -ddd option
When the debug option is given more than 2 times, it prints dynamic
index, function address and symbol name.  But it incorrectly assumed the
address is 64-bit long, but it actually is 'unsigned long' type.  So on
ARM, it wrongly consumes second argument for the first, and then
accessed the symbol table with a garbage index.  Boom..

Signed-off-by: Namhyung Kim <namhyung@gmail.com>