record: Fix daemon process ppid handling
Traditionally daemon processes have ppid of 1, so we catch those cases
to find a daemon process in the tid list.  However, AFAIK with
systemd (user daemon) it doesn't guaranteed anymore, so just treat tasks
that have ppid not set as daemon and set original ppid as pid (for
finding session info).

Signed-off-by: Namhyung Kim <namhyung@gmail.com>