record: Add --no-pltbind option
The --no-pltbind option is to set LD_BIND_NOT environment variable in
the child in order to trace concurrent library function calls.  The PLT
hooking is done by overwriting the function address in return path of
the first call.

However if the function blocks in the middle, concurrent accesses will
be missing since the dynamic resolver changes the function address in
GOT.  The PLT hooker overwrite the address again in the return path but
before the return it lose the accesses.  The glibc provides LD_BIND_NOT
option to cure this kind of situation but it requires the resolver keep
called everytime.

This option is not the default since it hurts the performance
significantly, so use with care.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>