record: Add -r/--threshold option
The --threshold option is to filter functions by time.  Functions which
their duration is shorter than the threshold will not be recorded.

This is done by the delayed write of rstacks which records the rstack at
function exit time.  At the function exit, it compares the execution
time to the threshold and skips writing small functions.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>