mcount: Fix trace-on/off wrt the delayed write
When the delayed write is used, trace-on/off feature can be misbehave.
Because writing is done on the exit path, earlier entry records might
not be recorded when exit is after the trace-off.  So flush existing
entry records if trace-off trigger is used.

Likewise, when it turns on the tracing later, records before the trigger
should be hidden.  So add MCOUNT_FL_DISABLED flag and skip when writing.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>