record: Add -A/--argument option
The -A/--argument option is to set function arguments to be printed.
The argument spec is recorded in the info file for replay.

Example looks like below:

  $ ftrace -A 'foo@arg1' -A 'bar@arg2'

Above example records the first argument of 'foo' function and the
second argument of 'bar' function.  It's possible to record more than
one arguments (separated by comma) of a function.

  $ ftrace -A 'foo@arg1,arg2'

It's user's responsibility to pass correct argument spec of functions.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>