record: Record return value during trace
This records return value for given functions with -R/--retval option.
It shares the same ftrace_arg_spec data with argument information and
distinguishes whether the given arg spec is arguements or retval by
identifying idx number.

For arguments, idx indicates the sequence order of arguments but idx can
never be 0.  So idx 0 can be used for storing return value then it can
be read by skiping if idx it not 0 when iterating arg spec list.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>