chrome: Support chrome trace format in ftrace dump
This patch makes ftrace dump the recorded data in chrome trace format.
It would be much easier to see the entire result in the bird's eye view
by using chrome trace tool, which is already shipped in chrome browser.

The usage is as follows:

  $ ftrace dump --chrome > trace_chrome_ftrace.json

You can go to 'chrome://tracing' page in chrome browser, and load the
output json data file that is generated by the above command to easily
see the ftrace result.

Basic keys for navigation in browser:
  w/s    Zoom in/out (+shift: faster)
  a/d    Pan left/right (+shift: faster)

Link: https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/preview#
Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>