mcount: Enable tracing time filtered functions
Time filter, '-r/--threshold' option, is useful to see the big picture
of execution sequence by pruning small functions but sometime we need to
see some small but important functions to understand the execution flow
better.

Trigger option '-T' can be used in this case, even if some functions
took less time than predefined threshold with '-r/--threshold' option,
ftrace will record the execution flow of functions that are marked with
-T '<func>@trace'

The usage is as follows:

  $ ftrace -r 100ms -T small_func@trace ./prog

In this case, all the functions that took less than 100ms will not be
recorded.  However, the function 'small_func' that is specified with
'trace' trigger can be recorded with the exact execution path.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>