mcount: Refactor record_trace_data() function call
When handling arguments and retval, record_trace_data() has to be
invoked with proper 'args_spec' data.

In order to record argument data, it has to be guaranteed that 'regs'
has to be passed with 'args_spec' that is not empty.

Likewise, in order to record return value, it has to be guaranteed that
'retval' has to be passed with 'args_spec' that is not empty.

These condition can be perfectly checked, but it increases code
complexity.  So this patch simplifies the check routine by wathcing only
'regs' or 'retval' value to decide if recorded routine has more field or
not.

  int record_trace_data(struct mcount_thread_data *mtdp,
  		      struct mcount_ret_stack *mrstack,
  		      struct list_head *args_spec,
  		      struct mcount_regs *regs,
  		      long *retval)

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>
[Fix segfault when force flush on fork/exec/...]
Signed-off-by: Namhyung Kim <namhyung@gmail.com>