symbol: Skip duplicate symbols
Honggyu reported that d8 built with clang sometimes crashed on replay.
It was because it had some duplicated symbols and found no args with the
symbols.  Such duplicated symbols recorded in the symbol file and then
had 0 size in the end since the symbol file doesn't have size field in
it,  (It just calculate the difference between two adjacent symbols and
use it as size).  Skip those duplicate symbols fixes the problem.

This patch removes artificial '__dynsym_end' symbol as it's duplicated
to the first normal symbol.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>