replay: Show small negative integers naturally
The default argument spec is 'long' but it's more common to use integer
in the function arguments.  Positive small integers are printed as
decimal in this case, but small negative interges are not since they're
zero-extended so treated as big positive numbers.  Add a check to print
them as decimal.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>