record: Fix a bug when buffers from same task written concurrently
With multi-threaded recording work, now each shmem_buffer can be written
by multiple threads concurrently.  However it's a problem when buffers
from same task are written concurrently.  As it's a race condition, the
recorded data might not have sequentially time-ordered record.

Add writer_list to track active writers and which task they are writing
the data.  Even if a buffer is available, do not process it when another
writer is working on the data from same task.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>