symbol: Use relative symbol address
In order to support dynamic libraries which can be loaded on an
arbitrary address, the symbol address should be relative.  The uftrace
replay (or other commands) should adjust the offset with the loaded
address in the map file.

This adds a new SYM_REL_ADDR feature bit to the file header to identify
if relative address is used.  This way we can handle both
older (absolute) address symbol file and new relative addresses.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>