demangle: Fix symbol length in the full demangler
It blindly assumed the max length of demangled symbol is 4 times of the
original length.  But there's a case it caused a segfault.

Actually, the __cxa_demangle() provides the symbol length in the third
argument to allocate proper size of memory.  So check the length before
the allocation.  It also fixes a possible memory leak when status is a
negative value.

Reported-by: Honggyu Kim <hong.gyu.kim@lge.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>