symbol: Factor out sort_dynsymtab()
The dynamic symbol table is a bit different in that it should reserve
the original ordering since it's checked by the index in plthook_entry()
using find_dynsym().  So instead of sorting by name, it uses sym_names
array to save the index.

Split out the code to sort_dynsymtab() so that loading symbol file code
can share the function.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>