record: Reduce kernel recording interval for -K
When full kernel tracing is enabled by -K/--kernel-full option, it
frequently shows LOST records due to kernel buffer overrun.  One thing
it can do is to wake writer threads more often and flush the buffer.
This patch changes it to 10us when -K is given, 1ms for -k and 1s for
normal tracing.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>