report: Fix cranch on out of range --sort-column
When showing diff mode report, --sort-column has to be 0, 1, or 2 for
base, target, or percentage respectively.  But if a user sets it to
other numbers, it just shows warning message and goes to normal
execution path, but it encounters assertion fail anyway as below:

  $ uftrace report --diff uftrace.data.old/ --sort-column 3
  invalid column number: 3
  uftrace: /home/honggyu/src/ftrace/cmd-report.c:278: cmp_diff_entry: Assertion `0' failed.
  Aborted (core dumped)

This is not a desirable usage so better to stop exection or proceed by
setting it to a specific column in this case.

In usual cases, user wants to sort based on diff percentage, so it'd be
better to fix --sort-colum to 2 by force then proceed normal execution
to show the report.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>