mcount: Restore return address on the higher depth
As node.js segfaults when it does garbage collection, there's a risk if
the target process accesses (and make use of) the hijacked return
addresses.  Although the uftrace provides 'recover' trigger to work
around this problem, it's hard for normal users to find the reason of
the crash and the exact location to apply the trigger.

This patch implements the automatic recover that is hook the return
address only the current level.  On the entry of a function, it hijacks
the return address of the current function and recover the return
address of the caller.  On the exit of the function, it again hijacks
the return address of the caller.

With this patch, it's now able to run the node.js without the trigger.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>