filter: Allow library name in the trigger action
Like "PLT" or "kernel", now library names can be specified in the
filter/trigger action.

  $ uftrace --force -F lib_b@libabc_test -- t-lib
  # DURATION    TID     FUNCTION
              [19048] | lib_b() {
     2.811 us [19048] |   lib_c();
     4.975 us [19048] | } /* lib_b */

The lib_a (now filtered out), lib_b and lib_c functions are in the
libabc_test_lib.so library.  The name was matched only using the given
prefix.  Users need to provided long names not to interfere with other
libraries.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>