mcount: Fix a bug regarding string and FP argument on x86
A string argument needs to save its length with the contents.  So it
calls strlen() during the save_to_argbuf() but implementation of
strlen() might use some SSE instructions which clobber XMM registers.
It resulted in invalid value for floating-point arguments if they used
together with string arguments.  Fix it by copying and calculating
length of string manually.

The same is true for memcpy in record_ret_stack().  And in this case,
the both buffers are aligned so that we can copy them in word unit.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>