symbol: Grow symbol table exponentially
When symbol table becomes very large, it suffers from the overhead
caused by realloc().  This was because it's called too often, copying
the old contents to new location over and over.  Instead of fixed
growing constant of 16, make it bigger as the table grows.

This reduces the time to load symbol table in chrome from 4 minutes to 4
seconds. :)

Signed-off-by: Namhyung Kim <namhyung@gmail.com>