chrome: Inform users when LOST records shown
Chrome trace format requires to have both entry and exit records so that
it can identify the range of function call and return.  However, if
there are some lost records, it cannot match the entry and exit of some
functions.  It may show some of functions do not return until the
program is finished or vice versa.

Since it's very difficult to generate fake records for lost data to
match entry and exit of some lost functions, we just inform the fact to
users as of now.

  $ uftrace dump -K --chrome > trace_kernel.json
  Some of function trace records are lost. (57 times shown)
  The output json format may not show the correct view in chrome browser.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>