kernel: Fallback to SyS_ symbol
The kallsyms uses the first symbol among those which have same
addresses.  This caused a problem when filtering syscall handlers - we
ususally use 'sys_*' name but it should be 'SyS_*' for most syscalls.
Provided a fallback when symbol starts with those prefix to convert
'SyS_' automatically (and compat_sys_ symbols as well).

With this patch, -F 'sys_close@kernel' filter works now.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>