test: Update filter-kernel test
The 'sys_get*@kernel' filter is improper for testing since it matches to
many kernel functions but in-kernel graph-filter size is fixed to 32.
The end result was that the sys_getegid() was not filtered due to the
size limit.

So update filter to only show sys_get-euid/egid functions.  This also
verifies recent change of the syscall name fallback mechanism.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>