report: Consider recursive calls
When recursive calls exist, the total time is accounted incorrectly.  In
the fibonacci example, main() calls fib() and it recursively calls
itself.  But uftrace report shows something like below:

  $ uftrace report -d fibonacci.data
    Total time   Self time  Nr. called  Function
    ==========  ==========  ==========  ====================================
     56.948 us   19.600 us           9  fib
     25.024 us    2.718 us           1  main
      2.853 us    2.853 us           1  __monstartup
      2.706 us    2.706 us           1  atoi
      2.194 us    2.194 us           1  __cxa_atexit

As you can see, the total time of fib is larger than the main's.  This
was because it adds total time at each recursive function call.  But we
need to omit the recursive time.

This patch checks function stack of current task, and if it found same
function address, mark current total time as recursive time.  When it
reports the total time it subtracts the recursive time.  So in the above
example, the total time and self time of fib() will be same.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>