replay: Add 'color' trigger action
The 'color' trigger is to highlight some specific functions in replay
output by showing them with given color.  This would be useful to easily
identify some functions from the rest of replay output.

The supported colors are
  red, green, blue, yellow, magenta, cyan, bold, and gray.

The simple usage is as follows:

  $ uftrace replay -T main@color=red

The above command highlights only 'main' function with 'red' color in
replay output.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>