record: Change ownership of data when sudo is used
Recording trace data with sudo created a directory with root privilege.
But subsequent record commands will fail because it cannot (re)move the
old data due to the permission error.  IMHO it'd be better to change
ownership of the data when creating it.  It also makes easier for other
commands to deal with the data.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>