dump: Add --sample-time option for flamegraph
The flamegraph works for sampling profilers so it basically counts the
number of samples.  But uftrace have full function call data, so passing
just number of calls would result in a different output.

Add --sample-time option is to control how it converts the time to
sample count.  It'll show more accurate result on the flamegraph.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>