build: Fix installation in user directory
The previous commit added bash completion support but it has a bug to
deal with the location of the installed file.  If user gives a
non-system directory is should preserve the prefix properly, but it just
removed it and try to use /etc instead which resulted in the 'permission
denied' error.

This patch also added support for configure script.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>