use "LINKFLAGS" for `ld` invocations
instead of "LDFLAGS", as that is supposed to be used for invocations of
gcc that are intended to call ld underneath, and not ld directly.
before this change, if a user overrode LDFLAGS for the main Makefile,
then it would also be overridden for these Makefiles as well, and would
often breaks (try, for example, `make LDFLAGS=-Wl,-z,now`).