kernel: Honor --kernel-only for replay and graph
The --kernel-only option was used only to dump kernel functions only.
It can be useful for replay and graph commands as well especially in
some systems that do not exactly match between kernel and user
timestamps.

In that case, we can only see kernel functions by ignoring user funtions
so that timestamps of user functions cannot confuse kernel function
record.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>