kernel: Fix kernel tracing on ARM
The kbuffer long size should be set correctly, but as pevent long size
is not set it didn't initialized properly and defaults to use 8-byte
with error.

Idealy, it needs to save the kernel tracing related data at record time,
and then use the info at analysis time.  But this patch is just a quick
fix using current system's long size until the real fix will be done.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>