graph: Fix depth filter before the function
When a depth filter set, it sometimes shows no result since the function
might be filtered out due to the depth filter.  For example, if we set
the depth of 2 to the basic t-abc program, the function 'c' cannot be
shown in the graph output.

In this case, user probably wanted to set the filter after the
function (not before).  So add a synthetic depth trigger for the
function to get the desired result.

Reported-by: Honggyu Kim <hong.gyu.kim@lge.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>