uftrace: Add -r/--time-range option
The --time-range option is to specify a range of time in order to limit
the scode of data analysis.  Instead of filtering by function name,
depth or duration, it could filter using timestamp of function.

The -r/--time-range option receives a start and/or a stop time separated
by '~'.  User can give any precision like below:

    $ uftrace replay -r 607.3346~ -f time,duration
  #     TIMESTAMP      DURATION    FUNCTION
        607.334643820            |   } /* wait */
        607.334644698            |   a() {
        607.334644854            |     b() {
        607.334644944            |       c() {
        607.334645167   1.234 us |         getpid();
        607.334646841   1.897 us |       } /* c */
        607.334647016   2.162 us |     } /* b */
        607.334647167   2.469 us |   } /* a */
        607.334647289            | } /* main */

Signed-off-by: Namhyung Kim <namhyung@gmail.com>