mcount: Prevent recursive call in mtd_dtor()
If the executable implements its own malloc and free, it could be called
during mtd_dtor() and access the rstack.  This is a use-after-free bug
and should be fixed by setting the recursion guard.  Note that it's not
needed to reset the guard variable since the thread is dead.

Another related bug is when the application uses pthread TSD API.
Because libmount also uses it to manage rstack and shmem buffer,
application should not access to them after it's destroyed.  But TSD
destructor is called as same order as registration, so libmcount is
called mostly first.  Thus other libraries calling a destructor can
enter libmcount and cause a segfault.

Reported-by: Honggyu Kim <hong.gyu.kim@lge.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>