record: Return an exit code according to child's exit status
It'd be great to return child's exit status so that other tools like
test runner can know the condition easily.  More detailed info can be
obtained by running uftrace info anyway.

The new enum was added to make it easy to identify the status and it
corresponds to usual exit status (SUCCESS and FAILURE).

Signed-off-by: Namhyung Kim <namhyung@gmail.com>