replay: Skip invalid remaining functions at the end
When some task ended in the middle of fstack, replay will show remaining
functions at the end.  But if the task is started during fork (or in a
time range) it might not set all previous stack with valid functions.
They'll be printed as 0 address but it'd be better to skip them.

Before:
  uftrace stopped tracing with remaining functions
  ===============================================
  task: 4315
  [14] <80bdfa0>
  [13] <80bdf20>
  [12] <21c80b0>
  [11] <0>
  [10] <0>
  [9] <0>
  [8] <0>
  [7] <0>
  [6] <0>
  [5] <0>
  [4] <0>
  [3] <0>
  [2] <0>
  [1] <0>
  [0] <0>

After:
  uftrace stopped tracing with remaining functions
  ===============================================
  task: 4315
  [2] <80bdfa0>
  [1] <80bdf20>
  [0] <21c80b0>

Reported-by: Honggyu Kim <hong.gyu.kim@lge.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>