mcount: Hook C++ exception handling functions
The __cxa_throw() function is called during C++ exception handling.
When the handling is done, we need to find how much stack frame is
unwinded but it's tricky since the __cxa_throw() never returns.

Instead, we can use __cxa_end_catch() which is called at the last stage
of exception handling.  By comparing the parent location and frame base
address, we can guess the number of unwinded frames and adjust the index
accordingly.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>