mcount: Save full rstack info for vfork
When vfork() is called all rstack info will be used by child for a
while.  As the child exited from the vfork, it reused the rstack for its
functions overwriting the parent's.  So save the current rstack info and
then restore when returning to parent.

Also change vforkexec test result to match to a correct output.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>