live: Show replay even in child terminated abnormally
The commit 532228e03b9c ("live: Use return value of record") changed to
run replay only if record returns UFTRACE_EXIT_SUCCESS.  But record
returns a value depending on the exit code of a child.  And it won't
show the result when child exits by a signal or returns a non-zero value.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>