session: Do not abort when task has no session
The session is used mostly for finding symbols from address.  Now
uftrace handles addresses have no matching symbol, it's better not
aborting on NULL session.

This is especially important because those data could not be used to run
any uftrace subcommand including "info" as well - make it hard to debug.

Normally a task should have at least a single session.  But sometimes it
didn't for some reason (maybe due to an order of recording the task
file).  So It added a debug message instead to help finding the cause
why it did not.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>