filter: Use kernel symbol table for filters
The setup_module_and_trigger() sets a symbol table for a given filter or
trigger spec.  But it always set NULL to a pointer to symbol table which
led to missing kernel filters or triggers even in replay time.

Use get_kernel_symtab() to set a loaded kernel symbol table if exists.
This fixes the following error:

  $ sudo uftrace -K3 -T ^sys_@kernel,depth=1 openclose
  fstack: /home/namhyung/project/uftrace/utils/fstack.c:342:fstack_setup_filters
    ERROR: failed to set filter or trigger: ^sys_@kernel,depth=1

Signed-off-by: Namhyung Kim <namhyung@gmail.com>