test: Fixup longjmp to __longjmp_chk
When -D_FORTIFY_SOURCE=2 is used (it seems the default setting of
ubuntu) the longjmp call is converted to __longjmp_chk.  So rename the
result to match with the actual output.

However it still has a problem causing a segfault.  It comes from
another __longjmp_chk() outside of the main().  Need to investigate
more.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>