record: Use poll to wait available data in writer threads
With kernel tracing, a writer thread should wait data from multiple
sources.  It'd be much efficient to wait using poll instead of a loop
with arbitrary timeouts.

To do this, it changes to use a pipe to signal the writer threads
rather than the pthread conditional variable.  By using a fd, it's easy
to wait on the buffer using poll.

This change reduces number of LOST messages in kernel significantly
while slightly increase kernel data due to IPI for waking up tasks.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>