mcount: Do not set bitfield directly
For performance and portability reasons, it'd be better to avoid using
bitfields.  So construct the struct manually.  I used ADDITION instead of
OR because it gave me a (tiny) speedup on my system.  Generated code
were identical except the add/or instructions.  Also using a temporary
variable reduced memory references.

This change saves ~3% of executation time when measuring a simple test
program with a tight loop.  I recorded it with perf and verified the
overhead in the record_ret_stack() was recuded significantly.

  $ perf diff
  # Event 'cycles:pp'
  #
  # Baseline  Delta Abs  Shared Object        Symbol
  # ........  .........  ...................  ............................
  #
      22.13%    -12.53%  libmcount-fast.so    [.] record_ret_stack
      17.62%     +0.34%  [vdso]               [.] 0x0000000000000930
       8.90%     +3.46%  [vdso]               [.] __vdso_clock_gettime
       5.57%     +0.45%  libmcount-fast.so    [.] mcount_entry
       4.36%     +1.93%  libmcount-fast.so    [.] mcount_return
       4.35%     +0.37%  [kernel.kallsyms]    [k] copy_user_generic_string
       4.32%     -0.16%  libmcount-fast.so    [.] record_trace_data
       3.64%     +0.67%  libmcount-fast.so    [.] mcount_exit
       2.82%     +0.48%  libpthread-2.24.so   [.] pthread_getspecific
       2.47%     +0.12%  libc-2.24.so         [.] __clock_gettime

Signed-off-by: Namhyung Kim <namhyung@gmail.com>